requirements.txt
discord.py
python-dotenv
duckduckgo-search
ollama
aiohttp